requirements.txt
pandas
streamlit
plotly
google-generativeai
python-dotenv
